app/requirements.txt
celery[redis]==5.4.0
Django==5.0.6
django-admin-sortable2==2.2.2
django-celery-beat==2.6.0
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-modeltranslation==0.19.7
djangogrpcframework==0.2.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
grpcio==1.47.0
grpcio-tools==1.47.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.18.1
